include a few general cleanup changes
git-svn-id: https://uncrustify.svn.sourceforge.net/svnroot/uncrustify/trunk@1611 74908cc9-670d-0410-99d4-ca74d6b17880

+++ src/args.cpp
@@ -100,11 +100,11 @@ const char *Args::Params(const char *token, int& index)
       return(NULL);
    }
 
-   token_len = strlen(token);
+   token_len = (int)strlen(token);
 
    for (idx = index; idx < m_count; idx++)
    {
-      arg_len = strlen(m_values[idx]);
+      arg_len = (int)strlen(m_values[idx]);
 
       if ((arg_len >= token_len) &&
           (memcmp(token, m_values[idx], token_len) == 0))
